Reduced type binding code
