+++ Code/EnginePlugins/TypeScriptPlugin/Components/TypeScriptComponent.cpp
@@ -145,93 +145,8 @@ void ezTypeScriptComponent::SetExposedVariables()
   {
     const auto& pair = m_Parameters.GetPair(p);
 
-    switch (pair.value.GetType())
-    {
-      case ezVariant::Type::Angle:
-        duk.SetNumberProperty(pair.key.GetString(), pair.value.Get<ezAngle>().GetRadian());
-        break;
-
-      case ezVariant::Type::Time:
-        duk.SetNumberProperty(pair.key.GetString(), pair.value.Get<ezTime>().GetSeconds());
-        break;
-
-      case ezVariant::Type::Bool:
-        duk.SetBoolProperty(pair.key.GetString(), pair.value.Get<bool>());
-        break;
-
-      case ezVariant::Type::Int8:
-      case ezVariant::Type::UInt8:
-      case ezVariant::Type::Int16:
-      case ezVariant::Type::UInt16:
-      case ezVariant::Type::Int32:
-      case ezVariant::Type::UInt32:
-      case ezVariant::Type::Int64:
-      case ezVariant::Type::UInt64:
-      case ezVariant::Type::Float:
-      case ezVariant::Type::Double:
-        duk.SetNumberProperty(pair.key.GetString(), pair.value.ConvertTo<double>());
-        break;
-
-      case ezVariant::Type::Color:
-        ezTypeScriptBinding::PushColor(duk, pair.value.Get<ezColor>());
-        duk.SetCustomProperty(pair.key.GetString());
-        break;
-
-      case ezVariant::Type::ColorGamma:
-        ezTypeScriptBinding::PushColor(duk, pair.value.Get<ezColorGammaUB>());
-        duk.SetCustomProperty(pair.key.GetString());
-        break;
-
-      case ezVariant::Type::Vector2:
-        ezTypeScriptBinding::PushVec2(duk, pair.value.Get<ezVec2>());
-        duk.SetCustomProperty(pair.key.GetString());
-        break;
-
-      case ezVariant::Type::Vector3:
-        ezTypeScriptBinding::PushVec3(duk, pair.value.Get<ezVec3>());
-        duk.SetCustomProperty(pair.key.GetString());
-        break;
-
-      case ezVariant::Type::Quaternion:
-        ezTypeScriptBinding::PushQuat(duk, pair.value.Get<ezQuat>());
-        duk.SetCustomProperty(pair.key.GetString());
-        break;
-
-      case ezVariant::Type::Transform:
-        ezTypeScriptBinding::PushTransform(duk, pair.value.Get<ezTransform>());
-        duk.SetCustomProperty(pair.key.GetString());
-        break;
-
-      case ezVariant::Type::String:
-        duk.SetStringProperty(pair.key.GetString(), pair.value.Get<ezString>());
-        break;
-
-      case ezVariant::Type::StringView:
-      {
-        ezStringBuilder tmp;
-
-        duk.SetStringProperty(pair.key.GetString(), pair.value.Get<ezStringView>().GetData(tmp));
-
-        break;
-      }
-
-        //case ezVariant::Type::Vector4:
-        //case ezVariant::Type::Vector2I:
-        //case ezVariant::Type::Vector3I:
-        //case ezVariant::Type::Vector4I:
-        //case ezVariant::Type::Vector2U:
-        //case ezVariant::Type::Vector3U:
-        //case ezVariant::Type::Vector4U:
-        //case ezVariant::Type::Matrix3:
-        //case ezVariant::Type::Matrix4:
-        //case ezVariant::Type::Uuid:
-        // TODO: implement these types
-
-      default:
-        EZ_ASSERT_NOT_IMPLEMENTED;
-    }
+    ezTypeScriptBinding::SetVariantProperty(duk, pair.key.GetString(), -1, pair.value); // [ comp ]
   }
-
 
   duk.PopStack(); // [ ]
 }
