+++ Code/EnginePlugins/TypeScriptPlugin/TsBinding/TsMessages.cpp
@@ -237,131 +237,8 @@ ezUniquePtr<ezMessage> ezTypeScriptBinding::MessageFromParameter(duk_context* pD
 
       ezAbstractMemberProperty* pMember = static_cast<ezAbstractMemberProperty*>(pProp);
 
-      const ezVariant::Type::Enum type = pMember->GetSpecificType()->GetVariantType();
-      switch (type)
-      {
-        case ezVariant::Type::Invalid:
-          break;
-
-        case ezVariant::Type::Bool:
-        {
-          bool value = duk.GetBoolProperty(pMember->GetPropertyName(), false, iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::String:
-        {
-          ezString value = duk.GetStringProperty(pMember->GetPropertyName(), "", iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::StringView:
-        {
-          ezStringView value = duk.GetStringProperty(pMember->GetPropertyName(), "", iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Int8:
-        case ezVariant::Type::Int16:
-        case ezVariant::Type::Int32:
-        case ezVariant::Type::Int64:
-        {
-          ezInt32 value = duk.GetIntProperty(pMember->GetPropertyName(), 0, iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::UInt8:
-        case ezVariant::Type::UInt16:
-        case ezVariant::Type::UInt32:
-        case ezVariant::Type::UInt64:
-        {
-          ezUInt32 value = duk.GetUIntProperty(pMember->GetPropertyName(), 0, iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Float:
-        {
-          const float value = duk.GetFloatProperty(pMember->GetPropertyName(), 0, iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Double:
-        {
-          const double value = duk.GetNumberProperty(pMember->GetPropertyName(), 0, iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Vector3:
-        {
-          ezVec3 value = ezTypeScriptBinding::GetVec3Property(duk, pMember->GetPropertyName(), iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Quaternion:
-        {
-          ezQuat value = ezTypeScriptBinding::GetQuatProperty(duk, pMember->GetPropertyName(), iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Transform:
-        {
-          ezTransform value = ezTypeScriptBinding::GetTransformProperty(duk, pMember->GetPropertyName(), iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Color:
-        {
-          ezColor value = ezTypeScriptBinding::GetColorProperty(duk, pMember->GetPropertyName(), iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::ColorGamma:
-        {
-          ezColorGammaUB value = ezTypeScriptBinding::GetColorProperty(duk, pMember->GetPropertyName(), iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Time:
-        {
-          const ezTime value = ezTime::Seconds(duk.GetNumberProperty(pMember->GetPropertyName(), 0, iObjIdx + 1));
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Angle:
-        {
-          const ezAngle value = ezAngle::Radian(duk.GetFloatProperty(pMember->GetPropertyName(), 0, iObjIdx + 1));
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Vector2:
-        {
-          ezVec2 value = ezTypeScriptBinding::GetVec2Property(duk, pMember->GetPropertyName(), iObjIdx + 1);
-          ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
-          break;
-        }
-
-        case ezVariant::Type::Matrix3:
-        case ezVariant::Type::Matrix4:
-        case ezVariant::Type::Uuid:
-        default:
-
-          EZ_ASSERT_NOT_IMPLEMENTED;
-          // TODO: implement these types
-      }
+      const ezVariant value = ezTypeScriptBinding::GetVariant(duk, iObjIdx + 1, pMember->GetSpecificType()->GetVariantType());
+      ezReflectionUtils::SetMemberPropertyValue(pMember, pMsg.Borrow(), value);
     }
   }
   else
@@ -420,87 +297,7 @@ void ezTypeScriptBinding::DukPutMessage(duk_context* pDuk, const ezMessage& msg)
 
     const ezVariant val = ezReflectionUtils::GetMemberPropertyValue(pMember, &msg);
 
-    const ezVariant::Type::Enum type = pMember->GetSpecificType()->GetVariantType();
-    switch (type)
-    {
-      case ezVariant::Type::Bool:
-      {
-        duk.SetBoolProperty(pMember->GetPropertyName(), val.Get<bool>());
-        break;
-      }
-
-      case ezVariant::Type::String:
-      case ezVariant::Type::StringView:
-      {
-        duk.SetStringProperty(pMember->GetPropertyName(), val.Get<ezString>());
-        break;
-      }
-
-      case ezVariant::Type::Int8:
-      case ezVariant::Type::Int16:
-      case ezVariant::Type::Int32:
-      case ezVariant::Type::Int64:
-      case ezVariant::Type::UInt8:
-      case ezVariant::Type::UInt16:
-      case ezVariant::Type::UInt32:
-      case ezVariant::Type::UInt64:
-      case ezVariant::Type::Float:
-      case ezVariant::Type::Double:
-      {
-        duk.SetNumberProperty(pMember->GetPropertyName(), val.ConvertTo<double>());
-        break;
-      }
-
-      case ezVariant::Type::Vector3:
-      {
-        SetVec3Property(duk, pMember->GetPropertyName(), -1, val.Get<ezVec3>());
-        break;
-      }
-
-      case ezVariant::Type::Color:
-      case ezVariant::Type::ColorGamma:
-      {
-        SetColorProperty(duk, pMember->GetPropertyName(), -1, val.ConvertTo<ezColor>());
-        break;
-      }
-
-      case ezVariant::Type::Time:
-      {
-        duk.SetNumberProperty(pMember->GetPropertyName(), val.Get<ezTime>().GetSeconds());
-        break;
-      }
-
-      case ezVariant::Type::Angle:
-      {
-        duk.SetNumberProperty(pMember->GetPropertyName(), val.Get<ezAngle>().GetRadian());
-        break;
-      }
-
-      case ezVariant::Type::Quaternion:
-      {
-        SetQuatProperty(duk, pMember->GetPropertyName(), -1, val.Get<ezQuat>());
-        break;
-      }
-
-      case ezVariant::Type::Vector2:
-      {
-        SetVec2Property(duk, pMember->GetPropertyName(), -1, val.Get<ezVec2>());
-        break;
-      }
-
-      case ezVariant::Type::Transform:
-      {
-        SetTransformProperty(duk, pMember->GetPropertyName(), -1, val.Get<ezTransform>());
-        break;
-      }
-
-      case ezVariant::Type::Matrix3:
-      case ezVariant::Type::Matrix4:
-      case ezVariant::Type::Uuid:
-      default:
-        // TODO: implement these types
-        EZ_ASSERT_NOT_IMPLEMENTED;
-    }
+    SetVariantProperty(duk, pMember->GetPropertyName(), -1, val);
   }
 }
